+++ src/climate_learn/transforms/__init__.py
@@ -1,0 +1,2 @@
+from .denormalize import Denormalize
+from .registry import TRANSFORMS_REGISTRY
